Move some attr utils from col_info to attr_utils
Move convert_none_to_empty_dict() and convert_single_to_list()
from models/collection_info.py to utils/attr_utils.py so
other modules can use them.
